refactor: sort the code out
